Define examples as part of pulsar-client-hs.cabal
This gives a bit more direct access to compiler and flags without having
to tweak the nix files. Builds are also faster as a change to the lib
doesn't lead to a whole rebuild of it in nix when the the examples
should be updated.

The drawback is that the examples are now part of the lib's nix build
result. However, that should be fine.