Do not bind to missing symbols / functions in C lib
The commented out functions refer to symbols that are not in the C
library.

This was revealed by the linker when the examples became part of the
Haskell lib's Cabal file. (Though, I don't get how the linking process
differs to what we had before. I checked that the symbols are really not
available with objdump.)